+++ lib/base/src/main/java/com/blankj/base/BaseApplication.java
@@ -8,7 +8,6 @@ import com.blankj.utilcode.util.AppUtils;
 import com.blankj.utilcode.util.CrashUtils;
 import com.blankj.utilcode.util.LogUtils;
 import com.blankj.utilcode.util.ProcessUtils;
-import com.squareup.leakcanary.LeakCanary;
 
 import java.util.ArrayList;
 
@@ -43,20 +42,8 @@ public class BaseApplication extends Application {
     public void onCreate() {
         super.onCreate();
         sInstance = this;
-        initLeakCanary();
         initLog();
         initCrash();
-    }
-
-    private void initLeakCanary() {// 内存泄露检查工具
-        if (isDebug()) {
-            if (LeakCanary.isInAnalyzerProcess(this)) {
-                // This process is dedicated to LeakCanary for heap analysis.
-                // You should not init your app in this process.
-                return;
-            }
-            LeakCanary.install(this);
-        }
     }
 
     // init it in ur application